requirements.txt
biopython==1.83
numpy==1.25.1
pandas==2.0.3
Requests==2.32.1
